Make size return an unsigned integer

+++ src/ManagedArray.hpp
@@ -75,7 +75,7 @@ class ManagedArray {
    *
    * \return The number of elements in the array
    */
-  CHAI_HOST size_t size() const;
+  CHAI_HOST uint size() const;
 
   /*!
    * \brief Return reference to i-th element of the ManagedArray.
